+++ spec/component_spec.rb
@@ -54,6 +54,29 @@ module Ovto
         ret = ExampleApp::ParentComp.new(nil).do_render(state: state)
         expect(ret).to eq("foo is 1")
       end
+
+      it "can pass falsy value as attribute" do
+        class ExampleApp
+          class ParentComp < Component
+            def render
+              o 'div' do
+                o ChildComp, foo: true
+                o ChildComp, foo: false
+                o ChildComp, foo: nil
+              end
+            end
+          end
+
+          class ChildComp < Component
+            def render(foo:)
+              foo.inspect
+            end
+          end
+        end
+
+        ret = ExampleApp::ParentComp.new(nil).do_render(state: :dummy)
+        expect(`JSON.stringify(ret)`).to eq('{"nodeName":"div","attributes":{},"children":["true","false","nil"]}')
+      end
     end
   end
 end