+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Fixes
 
 - Skip rendering if app state is not changed by an action
 - `o "div.main", class: 'hovered'` should yield `<div class='main hovered'>`
+- Cannot pass falsy value when rendering a child component
 
 ## v0.1.0 2018-06-01